% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurise.R
\name{featurise}
\alias{featurise}
\title{Add features to list of phones}
\usage{
featurise(phlist)
}
\arguments{
\item{phlist}{A list of phones or the output of \code{phonetise()}.}
}
\value{
A tibble.
}
\description{
This function counts occurrences of phones and includes basic phonetic features.
}
\examples{
ipa <- c("ada", "buba", "kiki", "saʃa")
ip_ph <- phonetise(ipa)
featurise(ip_ph)

}
