% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonetise.R
\name{get_no_ipa}
\alias{get_no_ipa}
\title{Get non-IPA characters.}
\usage{
get_no_ipa(chars)
}
\arguments{
\item{chars}{A vector of characters.}
}
\value{
A vector.
}
\description{
Given a vector of characters, it returns those which are not part of the IPA.
}
\examples{
get_no_ipa(c("a", "\0283", ">"))

}
