% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipa_symbols}
\alias{ipa_symbols}
\title{List of IPA symbols}
\format{
A data frame with 143 rows and 12 variables:
\describe{
\item{IPA}{IPA symbol.}
\item{unicode}{Unicode code.}
\item{uni_name}{Unicode name.}
\item{ipa_name}{IPA name.}
\item{phon_type}{The phonetic type of the symbol.}
\item{type}{General character type (\code{consonant}, \code{vowel}, \code{diacritic}).}
\item{height_ipa}{Vowel openness.}
\item{height}{Vowel height.}
\item{backness}{Vowel backness.}
\item{rounding}{Vowel rounding.}
\item{voicing}{Consonant voicing.}
\item{place}{Consonant place of articulation.}
\item{manner}{Consonant manner of articulation.}
\item{lateral}{Is the consonant lateral?}
\item{sonorant}{Is the phone sonorant?}
}
}
\usage{
ipa_symbols
}
\description{
List of IPA symbols
}
\keyword{datasets}
