% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kl_swadesh}
\alias{kl_swadesh}
\title{Klingon Swadesh list}
\format{
A data frame with 195 rows and 4 variables:
\describe{
\item{id}{Swadesh list item number.}
\item{gloss}{English gloss.}
\item{translit}{Klingon transliteration.}
\item{ipa}{IPA transcription.}
}
}
\usage{
kl_swadesh
}
\description{
The Swadesh list in Klingon.
}
\keyword{datasets}
