% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{ph_search}
\alias{ph_search}
\title{Search phones}
\usage{
ph_search(phlist, phonex)
}
\arguments{
\item{phlist}{The output of \code{phonetise()}.}

\item{phonex}{A phonetic expression. Supported shorthands are \code{C} for
consonant, \code{V} for vowel, and \verb{#} for word boundary.}
}
\value{
A list.
}
\description{
Given a vector of phonetised strings, find phones.
}
\examples{
ipa <- c("pʰãkʰ", "tʰum̥", "ɛkʰɯ", "pun")
ph <- c("pʰ", "tʰ", "kʰ", "ã", "m̥")
ipa_ph <- phonetise(ipa, multi = ph)
ph_search(ipa_ph, "#CV")

# partial matches are also returned
ph_search(ipa_ph, "p")

# use regular expressions
ph_search(ipa_ph, "pʰ?V")
}
