% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonetise.R
\name{phonetise}
\alias{phonetise}
\title{Tokenise IPA strings}
\usage{
phonetise(strings, multichar)
}
\arguments{
\item{strings}{A character vector with a list of words in IPA.}

\item{multichar}{A character vector of one or more multi-character phones as
strings.}
}
\value{
A tibble.
}
\description{
\code{phonetise()} tokenises strings of IPA symbols (like phonetic transcriptions
of words) into individual "phones". The output is a tibble with two columns:
\code{ipa}, the IPA strings, and \code{ipa_token}, the tokenised IPA strings as a list
column.
}
\examples{
ipa <- c("pʰãk", "tʰum̥", "ɛkʰɯ")
ph <- c("pʰ", "tʰ", "kʰ", "ã", "m̥")

phonetise(ipa, ph)

}
