% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonetisr-package.R
\docType{package}
\name{phonetisr-package}
\alias{phonetisr}
\alias{phonetisr-package}
\title{phonetisr: A Naive IPA Tokeniser}
\description{
It provides users with functions to parse International Phonetic Alphabet transcriptions into phones for further processing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stefanocoretta.github.io/phonetisr/}
}

}
\author{
\strong{Maintainer}: Stefano Coretta \email{stefano.coretta@gmail.com} (\href{https://orcid.org/0000-0001-9627-5532}{ORCID})

}
\keyword{internal}
